requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
openai==1.59.7
supabase==2.11.0
pydantic==2.10.5
httpx==0.27.2
starlette==0.41.3
typing_extensions==4.12.2
